+++ src/main/java/com/yourcodereview/java01/eugenet1024/api/requestValidation/ErrorHandlerControllerAdvice.java
@@ -1,4 +1,4 @@
-package com.yourcodereview.java01.eugenet1024.api.validation;
+package com.yourcodereview.java01.eugenet1024.api.requestValidation;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.FieldError;
